requirements.txt
pandas==1.4.0
PyMySQL==1.0.2
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3
openpyxl==3.0.9